+++ src/app/services/alert.service.ts
@@ -4,7 +4,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { lastValueFrom } from 'rxjs';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AlertService {
   // important: you cannot use html tags e.g. <br> in the alert messages, because it will not be rendered correctly
@@ -12,17 +12,44 @@ export class AlertService {
 
   constructor(
     private readonly alertController: AlertController,
-    private readonly translate: TranslateService) { }
+    private readonly translate: TranslateService
+  ) {}
 
   async showErrorAlert(errorKey: string) {
     const errorMessage = await lastValueFrom(this.translate.get(errorKey));
-    const errorTitle = await lastValueFrom(this.translate.get('ERROR_ALERT_TITLE_ERROR'));
-    const OKButton = await lastValueFrom(this.translate.get('ERROR_ALERT_BUTTON_OK'));
+    const errorTitle = await lastValueFrom(
+      this.translate.get('ERROR_ALERT_TITLE_ERROR')
+    );
+    const OKButton = await lastValueFrom(
+      this.translate.get('ERROR_ALERT_BUTTON_OK')
+    );
     const alert = await this.alertController.create({
       header: errorTitle,
       message: errorMessage,
-      buttons: [OKButton]
+      buttons: [OKButton],
     });
+    await alert.present();
+  }
+
+  async showStoragePermissionError(): Promise<void> {
+    const alert = await this.alertController.create({
+      header: await lastValueFrom(
+        this.translate.get('ERROR_MESSAGE_SAVE_QR_HEADER')
+      ),
+      subHeader: await lastValueFrom(
+        this.translate.get('ERROR_MESSAGE_SAVE_QR_SUBHEADER')
+      ),
+      message: await lastValueFrom(
+        this.translate.get('ERROR_MESSAGE_SAVE_QR_MESSAGE')
+      ),
+      buttons: [
+        {
+          text: 'OK',
+          role: 'cancel',
+        },
+      ],
+    });
+
     await alert.present();
   }
 }